feat(alerting): add query optimizations for prometheus (#76015)
